Major refactoring to deal with non-float components (vertex joint indices), nothing works anymore, but compiles. Might revert to using floats for the indices, although the new code should be in the right direction.

+++ src/hashers.h
@@ -7,22 +7,22 @@ struct CollectionHashers
 {
 	std::size_t operator()(const gsl::span<int>& vec) const
 	{
-		return hash_value(reinterpret_span<uint32_t>(vec));
+		return hash_value(reinterpret_span<ushort>(vec));
 	}
 
 	std::size_t operator()(const gsl::span<float>& vec) const
 	{
-		return hash_value(reinterpret_span<uint32_t>(vec));
+		return hash_value(reinterpret_span<ushort>(vec));
 	}
 
 	std::size_t operator()(const std::vector<int>& vec) const
 	{
-		return hash_value(reinterpret_span<uint32_t>(span(vec)));
+		return hash_value(reinterpret_span<ushort>(span(vec)));
 	}
 
 	std::size_t operator()(const std::vector<float>& vec) const
 	{
-		return hash_value(reinterpret_span<uint32_t>(span(vec)));
+		return hash_value(reinterpret_span<ushort>(span(vec)));
 	}
 };
